fix: move Spark vectorized classes under org.lance
