+++ lance-spark-base_2.12/src/main/java/org/lance/spark/vectorized/FixedSizeListAccessor.java
@@ -11,10 +11,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.spark.sql.vectorized;
+package org.lance.spark.vectorized;
 
 import org.apache.arrow.vector.ValueVector;
 import org.apache.arrow.vector.complex.FixedSizeListVector;
+import org.apache.spark.sql.vectorized.ColumnarArray;
 
 /**
  * Accessor for FixedSizeListVector that provides array access for Spark. This wraps a
