+++ lance-spark-base_2.12/src/main/java/org/lance/spark/vectorized/LanceLargeArrayAccessor.java
@@ -11,30 +11,41 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.spark.sql.vectorized;
+package org.lance.spark.vectorized;
 
 import org.apache.arrow.vector.complex.LargeListVector;
+import org.apache.spark.sql.vectorized.ColumnarArray;
 
 /**
  * Accessor for LargeListVector (64-bit offset lists) that wraps element vectors in
  * LanceArrowColumnVector. This ensures that elements are properly handled by Lance-specific
  * accessors.
  */
-public class LanceLargeArrayAccessor extends ArrowColumnVector.ArrowVectorAccessor {
+public class LanceLargeArrayAccessor {
 
   private final LargeListVector accessor;
   private final LanceArrowColumnVector arrayData;
 
   public LanceLargeArrayAccessor(LargeListVector vector) {
-    super(vector);
     this.accessor = vector;
     this.arrayData = new LanceArrowColumnVector(vector.getDataVector());
   }
 
-  @Override
-  final ColumnarArray getArray(int rowId) {
+  public boolean isNullAt(int rowId) {
+    return this.accessor.isNull(rowId);
+  }
+
+  public int getNullCount() {
+    return this.accessor.getNullCount();
+  }
+
+  public ColumnarArray getArray(int rowId) {
     long start = accessor.getElementStartIndex(rowId);
     long end = accessor.getElementEndIndex(rowId);
     return new ColumnarArray(arrayData, (int) start, (int) (end - start));
+  }
+
+  public void close() {
+    this.accessor.close();
   }
 }
