+++ lance-spark-base_2.12/src/main/java/org/lance/spark/vectorized/LargeVarCharAccessor.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.spark.sql.vectorized;
+package org.lance.spark.vectorized;
 
 import org.apache.arrow.vector.LargeVarCharVector;
 import org.apache.spark.unsafe.types.UTF8String;
