+++ lance-spark-base_2.12/src/main/java/org/lance/spark/vectorized/UInt1Accessor.java
@@ -11,23 +11,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.spark.sql.vectorized;
+package org.lance.spark.vectorized;
 
-import org.apache.arrow.vector.UInt8Vector;
+import org.apache.arrow.vector.UInt1Vector;
 
 /**
- * Accessor for unsigned 64-bit integers (UInt8). Maps to Spark LongType (may overflow for values
- * &gt; Long.MAX_VALUE, but no better option).
+ * Accessor for unsigned 8-bit integers (UInt1). Maps to Spark ShortType (signed 16-bit can hold all
+ * UInt8 values 0-255).
  */
-public class UInt8Accessor {
-  private final UInt8Vector accessor;
+public class UInt1Accessor {
+  private final UInt1Vector accessor;
 
-  UInt8Accessor(UInt8Vector vector) {
+  UInt1Accessor(UInt1Vector vector) {
     this.accessor = vector;
   }
 
-  final long getLong(int rowId) {
-    return accessor.getObjectNoOverflow(rowId).longValueExact();
+  final short getShort(int rowId) {
+    return (short) Byte.toUnsignedInt(accessor.get(rowId));
   }
 
   final boolean isNullAt(int rowId) {
