+++ lance-spark-base_2.12/src/test/scala/org/apache/spark/sql/vectorized/LanceArrowColumnVectorSuite.scala
@@ -29,6 +29,7 @@ import org.apache.spark.sql.types._
 import org.apache.spark.sql.util.{ArrowUtils, LanceArrowUtils}
 import org.apache.spark.unsafe.types.UTF8String
 import org.lance.spark.LanceConstant
+import org.lance.spark.vectorized.LanceArrowColumnVector
 import org.scalatest.funsuite.AnyFunSuite
 
 class LanceArrowColumnVectorSuite extends AnyFunSuite {
